+++ ros2_unbag/exporters/__init__.py
@@ -3,13 +3,17 @@
 from .csv_exporter import export_topic_csv
 from .image_exporter import export_topic_images
 from .jsonl_exporter import export_topic_jsonl
+from .parquet_exporter import export_topic_parquet
 from .raw_exporter import export_topic_raw
+from .sqlite_exporter import export_topic_sqlite
 from .video_exporter import export_topic_video
 
 __all__ = [
     "export_topic_csv",
     "export_topic_images",
     "export_topic_jsonl",
+    "export_topic_parquet",
     "export_topic_raw",
+    "export_topic_sqlite",
     "export_topic_video",
 ]